Restructure MultiCouplingFourSolvers*.
Example implementing "every test has its own file":

* Move all tests and helpers related to MultiCouplingFourSolvers into folder src/precice/serial/tests/serial/multiCouplingFourSolvers
* Create individual .cpp files for MultiCouplingFourSolvers1 and MultiCouplingFourSolvers2.
* Introduce Test suite MultiCouplingFourSolvers containing MultiCouplingFourSolvers1 and MultiCouplingFourSolvers2.
* Move related configs into folder src/precice/serial/tests/serial/multiCouplingFourSolvers.
* Clean-up includes of all files.

Related to https://github.com/precice/precice/issues/1021#issuecomment-975191360.
