+++ src/tests.cmake
@@ -56,6 +56,9 @@ target_sources(testprecice
     src/precice/tests/DataContextTest.cpp
     src/precice/tests/ParallelTests.cpp
     src/precice/tests/SerialTests.cpp
+    src/precice/tests/serial/multiCouplingFourSolvers/helpers.cpp
+    src/precice/tests/serial/multiCouplingFourSolvers/MultiCouplingFourSolvers1.cpp
+    src/precice/tests/serial/multiCouplingFourSolvers/MultiCouplingFourSolvers2.cpp
     src/precice/tests/ToolingTests.cpp
     src/precice/tests/VersioningTests.cpp
     src/precice/tests/WatchIntegralTest.cpp
